+++ contrib/opencensus-ext-flask/opencensus/ext/flask/flask_middleware.py
@@ -176,12 +176,14 @@ class FlaskMiddleware(object):
 
         try:
             tracer = execution_context.get_opencensus_tracer()
-            tracer.add_attribute_to_current_span(
-                HTTP_ROUTE, flask.request.url_rule.rule
-            )
+            url_rule = flask.request.url_rule
+            if url_rule is not None:
+                tracer.add_attribute_to_current_span(
+                    HTTP_ROUTE, url_rule.rule
+                )
             tracer.add_attribute_to_current_span(
                 HTTP_STATUS_CODE,
-                str(response.status_code)
+                response.status_code
             )
         except Exception:  # pragma: NO COVER
             log.error('Failed to trace request', exc_info=True)