+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -5,6 +5,8 @@
   ([#767](https://github.com/census-instrumentation/opencensus-python/pull/767))
 - Standard metrics incoming requests per second
   ([#758](https://github.com/census-instrumentation/opencensus-python/pull/758))
+- Standard metrics incoming requests average execution rate
+  ([#773](https://github.com/census-instrumentation/opencensus-python/pull/773))
 
 ## 0.7.0
 Released 2019-07-31